+++ src/framework/theme/components/accordion/accordion-body.component.scss
@@ -4,7 +4,6 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
 
- :host {
-   display: flex;
-   flex-direction: column;
- }
+:host {
+  padding: 0 1.5rem 1rem;
+}
